% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat5_structure.R
\name{get_stat5_structure}
\alias{get_stat5_structure}
\title{Retrieve STAT5 Structure Using PDB ID}
\usage{
get_stat5_structure(pdb_id)
}
\arguments{
\item{pdb_id}{A character string containing the PDB ID of the STAT5 protein.}
}
\value{
A data structure containing the STAT5 protein structure.
}
\description{
This function retrieves a STAT5 protein structure using its PDB ID.
It requires the 'httr' and 'jsonlite' packages to send a request to the PDB API and parse the response.
}
\examples{
get_stat5_structure("PDB_ID_HERE")
}
\references{
Berman, Helen M., et al. The Protein Data Bank. Nucleic Acids Research, 2000.
}
