% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat5_structure.R
\name{search_stat5_structures}
\alias{search_stat5_structures}
\title{Search for STAT5 Structures in PDB}
\usage{
search_stat5_structures(query)
}
\arguments{
\item{query}{A character string containing the search query.}
}
\value{
A character vector of PDB IDs matching the query.
}
\description{
This function searches the Protein Data Bank for structures related to STAT5 using a given query string.
It uses the 'httr' package to send requests to the PDB API.
}
\examples{
search_stat5_structures("STAT5")
}
\references{
Berman, Helen M., et al. The Protein Data Bank. Nucleic Acids Research, 2000.
}
