% Generated by roxygen2: do not edit by hand
\name{show_stat5_structure}
\alias{show_stat5_structure}
\title{Show STAT5 Structure}
\description{
  Retrieves and displays the three-dimensional structure of the STAT5 protein from the Protein Data Bank (PDB).
}
\usage{
show_stat5_structure(pdb_id, display_mode = "default")
}
\arguments{
  \item{pdb_id}{A character string specifying the PDB ID of the STAT5 protein.}
  \item{display_mode}{An optional parameter to specify the mode of structure display, such as "ribbon", "wireframe", etc.}
}
\value{
  A plot or graphical representation of the STAT5 structure.
}
\examples{
  show_stat5_structure("1Y1U") # Example PDB ID
  show_stat5_structure("1Y1U", display_mode = "ribbon")
}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023).
}
