% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show Expression Plot.R
\name{show_expression_plot}
\alias{show_expression_plot}
\title{Show Expression Plot}
\description{
  Generates a bar plot for protein expression levels.
}
\usage{
show_expression_plot(expression_data)
}
\arguments{
  \item{expression_data}{A data frame with two columns: Proteins and ExpressionLevels.}
}
\value{
  A bar plot showing the expression levels of proteins.
}
\examples{
  proteins <- c("Protein A", "Protein B", "Protein C")
  levels <- c(23, 45, 10)
  data <- data.frame(Proteins = proteins, ExpressionLevels = levels)
  show_expression_plot(data)
}
\references{
H. Wickham
